ajouter des fonctionnalitées pour abonnement

+++ backend/src/main/java/tn/kidora/spring/kidorabackoffice/dto/AbonnementRequestDTO.java
@@ -11,6 +11,6 @@ public class AbonnementRequestDTO {
     private Double montantPaye;
     private Double montantDu;
     private StatutPaiement statut;
-    private Long etablissementId;
+    private Integer etablissementId;
     
 }
